# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customer": {
      "post": {
        "operationId": "CustomerController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "添加客户",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "CustomerController_findAll",
        "parameters": [],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取客户列表",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customer/{id}": {
      "get": {
        "operationId": "CustomerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取客户详情",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新客户信息",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "CustomerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除客户",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/jobs": {
      "post": {
        "operationId": "JobController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建新任务",
        "tags": [
          "Jobs"
        ]
      },
      "get": {
        "operationId": "JobController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Current page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "按状态筛选: PENDING, IN_PROGRESS, COMPLETED, ISSUE, CANCELLED",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "按客户ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workerId",
            "required": false,
            "in": "query",
            "description": "按员工ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "开始日期 (ISO 8601)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "结束日期 (ISO 8601)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取任务列表（支持筛选和分页）",
        "tags": [
          "Jobs"
        ]
      }
    },
...
    "/invoice/generate/{jobId}": {
      "post": {
        "operationId": "InvoiceController_generateFromJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "从已完成任务自动生成账单",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoice": {
      "post": {
        "operationId": "InvoiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "手动创建账单",
        "tags": [
          "Invoices"
        ]
      },
      "get": {
        "operationId": "InvoiceController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "Current page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "UNPAID, PAID, VOID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取账单列表",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoice/{id}": {
      "get": {
        "operationId": "InvoiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取账单详情",
        "tags": [
          "Invoices"
        ]
      },
      "patch": {
        "operationId": "InvoiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新账单（仅限未支付状态）",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoice/{id}/pay": {
      "post": {
        "operationId": "InvoiceController_generatePaymentLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "生成 Stripe 支付链接",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoice/{id}/mark-paid": {
      "post": {
        "operationId": "InvoiceController_markAsPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "手动标记已支付",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/invoice/{id}/void": {
      "post": {
        "operationId": "InvoiceController_voidInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "未登录或 Token 无效"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "作废账单",
        "tags": [
          "Invoices"
        ]
      }
    },
...
  },
  "info": {
    "title": "CleanOps API",
    "description": "The CleanOps 2026 Management API - 专业清洁业务管理系统",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice Johnson"
          },
          "address": {
            "type": "string",
            "example": "123 Dublin St, Dublin 2",
            "description": "完整地址"
          },
          "eircode": {
            "type": "string",
            "example": "D02 X123",
            "description": "爱尔兰 Eircode"
          },
          "accessCode": {
            "type": "string",
            "example": "Key under the mat / Code: 1234",
            "description": "入户方式"
          },
          "lat": {
            "type": "number",
            "example": 53.3498
          },
          "lng": {
            "type": "number",
            "example": -6.2603
          },
          "companyId": {
            "type": "string",
            "example": "uuid-of-company"
          },
          "isCommercial": {
            "type": "boolean",
            "example": true,
            "description": "商业/住宅"
          }
        },
        "required": [
          "name",
          "address",
          "eircode",
          "accessCode",
          "lat",
          "lng"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice Johnson"
          },
          "address": {
            "type": "string",
            "example": "123 Dublin St, Dublin 2",
            "description": "完整地址"
          },
          "eircode": {
            "type": "string",
            "example": "D02 X123",
            "description": "爱尔兰 Eircode"
          },
          "accessCode": {
            "type": "string",
            "example": "Key under the mat / Code: 1234",
            "description": "入户方式"
          },
          "lat": {
            "type": "number",
            "example": 53.3498
          },
          "lng": {
            "type": "number",
            "example": -6.2603
          },
          "companyId": {
            "type": "string",
            "example": "uuid-of-company"
          },
          "isCommercial": {
            "type": "boolean",
            "example": true,
            "description": "商业/住宅"
          }
        }
      },
...
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "scheduledStart": {
            "type": "string",
            "example": "2026-05-01T09:00:00Z"
          },
          "estimatedDuration": {
            "type": "number",
            "example": 120,
            "description": "预计耗时（分钟）"
          },
          "customerId": {
            "type": "string",
            "example": "customer-uuid"
          },
          "workerIds": {
            "example": [
              "worker-uuid-1"
            ],
            "description": "指派的员工ID列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notes": {
            "type": "string",
            "example": "请带上备用钥匙"
          }
        },
        "required": [
          "scheduledStart",
          "customerId"
        ]
      },
...
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "example": "job-uuid",
            "description": "关联任务ID"
          },
          "amount": {
            "type": "number",
            "example": 5000,
            "description": "总金额（分）"
          },
          "vatAmount": {
            "type": "number",
            "example": 750,
            "description": "VAT 税额（分）"
          },
          "pensionAmount": {
            "type": "number",
            "example": 75,
            "description": "养老金提留额（分）",
            "default": 0
          },
          "companyId": {
            "type": "string",
            "example": "company-uuid"
          }
        },
        "required": [
          "jobId",
          "amount",
          "vatAmount",
          "companyId"
        ]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "vatAmount": {
            "type": "number"
          },
          "pensionAmount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "description": "UNPAID, PAID, VOID"
          }
        }
      }
    }
  }
}